refactor(agent-toolkit): reorganize agents tools into per-tool folders
Restructure agents-tools directory with each tool in its own folder alongside its tests for better clarity and organization.

New structure:
- create-agent/
  - create-agent-tool.ts
  - create-agent-tool.test.ts
- get-agent/
  - get-agent-tool.ts
  - get-agent-tool.test.ts
- delete-agent/
  - delete-agent-tool.ts
  - delete-agent-tool.test.ts
- shared/
  - agents.graphql.dev.ts

All tests passing. No behavior changes.
