+++ packages/agent-toolkit/src/core/tools/platform-api-tools/agents-tools/delete-agent/delete-agent-tool.test.ts
@@ -1,5 +1,5 @@
 import { MondayAgentToolkit } from 'src/mcp/toolkit';
-import { callToolByNameRawAsync, createMockApiClient, parseToolResult } from '../test-utils/mock-api-client';
+import { callToolByNameRawAsync, createMockApiClient, parseToolResult } from '../../test-utils/mock-api-client';
 import { DeleteAgentMutation } from 'src/monday-graphql/generated/graphql.dev/graphql';
 
 describe('DeleteAgentTool', () => {
