+++ packages/agent-toolkit/src/core/tools/platform-api-tools/agents-tools/delete-agent/delete-agent-tool.ts
@@ -2,11 +2,11 @@ import { z } from 'zod';
 import {
   DeleteAgentMutation,
   DeleteAgentMutationVariables,
-} from '../../../../monday-graphql/generated/graphql.dev/graphql';
-import { deleteAgentMutation } from './agents.graphql.dev';
-import { ToolInputType, ToolOutputType, ToolType } from '../../../tool';
-import { BaseMondayApiTool, createMondayApiAnnotations } from '../base-monday-api-tool';
-import { rethrowWithContext } from '../../../../utils';
+} from '../../../../../monday-graphql/generated/graphql.dev/graphql';
+import { deleteAgentMutation } from '../shared/agents.graphql.dev';
+import { ToolInputType, ToolOutputType, ToolType } from '../../../../tool';
+import { BaseMondayApiTool, createMondayApiAnnotations } from '../../base-monday-api-tool';
+import { rethrowWithContext } from '../../../../../utils';
 
 export const deleteAgentToolSchema = {
   id: z
