+++ packages/agent-toolkit/src/core/tools/platform-api-tools/agents-tools/index.ts
@@ -1,3 +1,3 @@
-export * from './get-agent-tool';
-export * from './create-agent-tool';
-export * from './delete-agent-tool';
+export * from './get-agent/get-agent-tool';
+export * from './create-agent/create-agent-tool';
+export * from './delete-agent/delete-agent-tool';
